Scene 클래스들 Scenes 파일로 통합
